+++ federation/callback.ts
@@ -18,6 +18,11 @@ export type NodeInfoDispatcher<TContextData> = (
  * A callback that dispatches an {@link Actor} object.
  *
  * @typeParam TContextData The context data to pass to the {@link Context}.
+ * @param context The request context.
+ * @param handle The actor's handle.
+ * @param key The public key of the actor.  **Deprecated: Do not rely on this
+ *            parameter. Instead, use the {@link Context.getActorKeyPairs}
+ *            method.**
  */
 export type ActorDispatcher<TContextData> = (
   context: RequestContext<TContextData>,
@@ -26,9 +31,24 @@ export type ActorDispatcher<TContextData> = (
 ) => Actor | null | Promise<Actor | null>;
 
 /**
+ * A callback that dispatches key pairs for an actor.
+ *
+ * @typeParam TContextData The context data to pass to the {@link Context}.
+ * @param contextData The context data.
+ * @param handle The actor's handle.
+ * @returns The key pairs.
+ * @since 0.10.0
+ */
+export type ActorKeyPairsDispatcher<TContextData> = (
+  contextData: TContextData,
+  handle: string,
+) => CryptoKeyPair[] | Promise<CryptoKeyPair[]>;
+
+/**
  * A callback that dispatches a key pair for an actor.
  *
  * @typeParam TContextData The context data to pass to the {@link Context}.
+ * @deprecated
  */
 export type ActorKeyPairDispatcher<TContextData> = (
   contextData: TContextData,