+++ runtime/key.ts
@@ -1,34 +1,57 @@
 import { decodeBase64, encodeBase64 } from "@std/encoding/base64";
+import { PublicKeyInfo } from "pkijs";
+import { validateCryptoKey } from "../sig/key.ts";
+
+const algorithms: Record<
+  string,
+  | AlgorithmIdentifier
+  | HmacImportParams
+  | RsaHashedImportParams
+  | EcKeyImportParams
+> = {
+  "1.2.840.113549.1.1.1": { name: "RSASSA-PKCS1-v1_5", hash: "SHA-256" },
+  "1.3.101.112": "Ed25519",
+};
 
 /**
- * Imports a PEM-SPKI formatted RSA-PKCS#1-v1.5 public key.
- * @param pem The PEM-SPKI formatted RSA-PKCS#1-v1.5 public key.
+ * Imports a PEM-SPKI formatted public key.
+ * @param pem The PEM-SPKI formatted public key.
  * @returns The imported public key.
+ * @throws {TypeError} If the key is invalid or unsupported.
  * @since 0.5.0
  */
 export async function importSpki(pem: string): Promise<CryptoKey> {
   pem = pem.replace(/(?:-----(?:BEGIN|END) PUBLIC KEY-----|\s)/g, "");
-  const spki = decodeBase64(pem);
+  let spki: Uint8Array;
+  try {
+    spki = decodeBase64(pem);
+  } catch (_) {
+    throw new TypeError("Invalid PEM-SPKI format.");
+  }
+  const pki = PublicKeyInfo.fromBER(spki);
+  const oid = pki.algorithm.algorithmId;
+  const algorithm = algorithms[oid];
+  if (algorithm == null) {
+    throw new TypeError("Unsupported algorithm: " + oid);
+  }
   return await crypto.subtle.importKey(
     "spki",
     spki,
-    { name: "RSASSA-PKCS1-v1_5", hash: "SHA-256" },
+    algorithm,
     true,
     ["verify"],
   );
 }
 
 /**
- * Exports an RSA-PKCS#1-v1.5 public key in PEM-SPKI format.
+ * Exports a public key in PEM-SPKI format.
  * @param key The public key to export.
- * @returns The exported RSA-PKCS#1-v1.5 public key in PEM-SPKI format.
- * @since 0.5.0
+ * @returns The exported public key in PEM-SPKI format.
  * @throws {TypeError} If the key is invalid or unsupported.
+ * @since 0.5.0
  */
 export async function exportSpki(key: CryptoKey): Promise<string> {
-  if (key.algorithm.name !== "RSASSA-PKCS1-v1_5") {
-    throw new TypeError("Unsupported algorithm: " + key.algorithm.name);
-  }
+  validateCryptoKey(key);
   const spki = await crypto.subtle.exportKey("spki", key);
   let pem = encodeBase64(spki);
   pem = (pem.match(/.{1,64}/g) || []).join("\n");