+++ testing/context.ts
@@ -2,6 +2,7 @@ import { getLogger } from "@logtape/logtape";
 import type { Context, RequestContext } from "../federation/context.ts";
 import { RouterError } from "../federation/router.ts";
 import { mockDocumentLoader } from "./docloader.ts";
+import { CryptographicKey } from "../vocab/vocab.ts";
 
 export function createContext<TContextData>(
   {
@@ -16,6 +17,7 @@ export function createContext<TContextData>(
     getFollowingUri,
     getFollowersUri,
     parseUri,
+    getActorKeyPairs,
     getActorKey,
     getDocumentLoader,
     sendActivity,
@@ -50,8 +52,28 @@ export function createContext<TContextData>(
     getDocumentLoader: getDocumentLoader ?? ((_params) => {
       throw new Error("Not implemented");
     }),
-    getActorKey: getActorKey ?? ((_handle) => {
-      return Promise.resolve(null);
+    getActorKeyPairs: getActorKeyPairs ?? ((_handle) => Promise.resolve([])),
+    getActorKey: getActorKey ?? (async (handle) => {
+      getLogger(["fedify", "federation"]).warn(
+        "Context.getActorKeys() method is deprecated; " +
+          "use Context.getActorKeyPairs() method instead.",
+      );
+      if (getActorKeyPairs == null) return null;
+      for (const keyPair of await getActorKeyPairs(handle)) {
+        const { privateKey } = keyPair;
+        if (
+          privateKey.algorithm.name !== "RSASSA-PKCS1-v1_5" ||
+          (privateKey.algorithm as unknown as { hash: { name: string } }).hash
+              .name !==
+            "SHA-256"
+        ) continue;
+        return new CryptographicKey({
+          id: keyPair.keyId,
+          owner: getActorUri?.(handle) ?? throwRouteError(),
+          publicKey: keyPair.publicKey,
+        });
+      }
+      return null;
     }),
     sendActivity: sendActivity ?? ((_params) => {
       throw new Error("Not implemented");